fix(tests): correct test-support import paths after package move
